Move importers into module
