+++ bookwyrm/importers/__init__.py
@@ -1,0 +1,5 @@
+""" import classes """
+
+from .importer import Importer
+from .goodreads_import import GoodreadsImporter
+from .librarything_import import LibrarythingImporter
