+++ bookwyrm/tests/importers/test_goodreads_import.py
@@ -1,4 +1,5 @@
 """ testing import """
+from collections import namedtuple
 import csv
 import pathlib
 from unittest.mock import patch
@@ -6,23 +7,21 @@ from unittest.mock import patch
 from django.test import TestCase
 import responses
 
-from bookwyrm import models, importer
-from bookwyrm.librarything_import import LibrarythingImporter
+from bookwyrm import models
+from bookwyrm.importers import importer, GoodreadsImporter
 from bookwyrm.settings import DOMAIN
 
 
-class LibrarythingImport(TestCase):
-    """ importing from librarything tsv """
+class GoodreadsImport(TestCase):
+    """ importing from goodreads csv """
 
     def setUp(self):
-        """ use a test tsv """
-        self.importer = LibrarythingImporter()
-        datafile = pathlib.Path(__file__).parent.joinpath("data/librarything.tsv")
-
-        # Librarything generates latin encoded exports...
+        """ use a test csv """
+        self.importer = GoodreadsImporter()
+        datafile = pathlib.Path(__file__).parent.joinpath("data/goodreads.csv")
         self.csv = open(datafile, "r", encoding=self.importer.encoding)
         self.user = models.User.objects.create_user(
-            "mmai", "mmai@mmai.mmai", "password", local=True
+            "mouse", "mouse@mouse.mouse", "password", local=True
         )
 
         models.Connector.objects.create(
@@ -53,11 +52,11 @@ class LibrarythingImport(TestCase):
         import_items = models.ImportItem.objects.filter(job=import_job).all()
         self.assertEqual(len(import_items), 3)
         self.assertEqual(import_items[0].index, 0)
-        self.assertEqual(import_items[0].data["Book Id"], "5498194")
+        self.assertEqual(import_items[0].data["Book Id"], "42036538")
         self.assertEqual(import_items[1].index, 1)
-        self.assertEqual(import_items[1].data["Book Id"], "5015319")
+        self.assertEqual(import_items[1].data["Book Id"], "52691223")
         self.assertEqual(import_items[2].index, 2)
-        self.assertEqual(import_items[2].data["Book Id"], "5015399")
+        self.assertEqual(import_items[2].data["Book Id"], "28694510")
 
     def test_create_retry_job(self):
         """ trying again with items that didn't import """
@@ -73,9 +72,20 @@ class LibrarythingImport(TestCase):
         retry_items = models.ImportItem.objects.filter(job=retry).all()
         self.assertEqual(len(retry_items), 2)
         self.assertEqual(retry_items[0].index, 0)
-        self.assertEqual(import_items[0].data["Book Id"], "5498194")
+        self.assertEqual(retry_items[0].data["Book Id"], "42036538")
         self.assertEqual(retry_items[1].index, 1)
-        self.assertEqual(retry_items[1].data["Book Id"], "5015319")
+        self.assertEqual(retry_items[1].data["Book Id"], "52691223")
+
+    def test_start_import(self):
+        """ begin loading books """
+        import_job = self.importer.create_job(self.user, self.csv, False, "unlisted")
+        MockTask = namedtuple("Task", ("id"))
+        mock_task = MockTask(7)
+        with patch("bookwyrm.importer.import_data.delay") as start:
+            start.return_value = mock_task
+            self.importer.start_import(import_job)
+        import_job.refresh_from_db()
+        self.assertEqual(import_job.task_id, "7")
 
     @responses.activate
     def test_import_data(self):
@@ -94,16 +104,14 @@ class LibrarythingImport(TestCase):
         self.assertEqual(import_item.book.id, book.id)
 
     def test_handle_imported_book(self):
-        """ librarything import added a book, this adds related connections """
+        """ goodreads import added a book, this adds related connections """
         shelf = self.user.shelf_set.filter(identifier="read").first()
         self.assertIsNone(shelf.books.first())
 
         import_job = models.ImportJob.objects.create(user=self.user)
-        datafile = pathlib.Path(__file__).parent.joinpath("data/librarything.tsv")
-        csv_file = open(datafile, "r", encoding=self.importer.encoding)
-        for index, entry in enumerate(
-            list(csv.DictReader(csv_file, delimiter=self.importer.delimiter))
-        ):
+        datafile = pathlib.Path(__file__).parent.joinpath("data/goodreads.csv")
+        csv_file = open(datafile, "r")
+        for index, entry in enumerate(list(csv.DictReader(csv_file))):
             entry = self.importer.parse_fields(entry)
             import_item = models.ImportItem.objects.create(
                 job_id=import_job.id, index=index, data=entry, book=self.book
@@ -121,25 +129,23 @@ class LibrarythingImport(TestCase):
         readthrough = models.ReadThrough.objects.get(user=self.user)
         self.assertEqual(readthrough.book, self.book)
         # I can't remember how to create dates and I don't want to look it up.
-        self.assertEqual(readthrough.start_date.year, 2007)
-        self.assertEqual(readthrough.start_date.month, 4)
-        self.assertEqual(readthrough.start_date.day, 16)
-        self.assertEqual(readthrough.finish_date.year, 2007)
-        self.assertEqual(readthrough.finish_date.month, 5)
-        self.assertEqual(readthrough.finish_date.day, 8)
+        self.assertEqual(readthrough.start_date.year, 2020)
+        self.assertEqual(readthrough.start_date.month, 10)
+        self.assertEqual(readthrough.start_date.day, 21)
+        self.assertEqual(readthrough.finish_date.year, 2020)
+        self.assertEqual(readthrough.finish_date.month, 10)
+        self.assertEqual(readthrough.finish_date.day, 25)
 
     def test_handle_imported_book_already_shelved(self):
-        """ librarything import added a book, this adds related connections """
+        """ goodreads import added a book, this adds related connections """
         with patch("bookwyrm.models.activitypub_mixin.broadcast_task.delay"):
             shelf = self.user.shelf_set.filter(identifier="to-read").first()
             models.ShelfBook.objects.create(shelf=shelf, user=self.user, book=self.book)
 
         import_job = models.ImportJob.objects.create(user=self.user)
-        datafile = pathlib.Path(__file__).parent.joinpath("data/librarything.tsv")
-        csv_file = open(datafile, "r", encoding=self.importer.encoding)
-        for index, entry in enumerate(
-            list(csv.DictReader(csv_file, delimiter=self.importer.delimiter))
-        ):
+        datafile = pathlib.Path(__file__).parent.joinpath("data/goodreads.csv")
+        csv_file = open(datafile, "r")
+        for index, entry in enumerate(list(csv.DictReader(csv_file))):
             entry = self.importer.parse_fields(entry)
             import_item = models.ImportItem.objects.create(
                 job_id=import_job.id, index=index, data=entry, book=self.book
@@ -156,22 +162,20 @@ class LibrarythingImport(TestCase):
         self.assertIsNone(self.user.shelf_set.get(identifier="read").books.first())
         readthrough = models.ReadThrough.objects.get(user=self.user)
         self.assertEqual(readthrough.book, self.book)
-        self.assertEqual(readthrough.start_date.year, 2007)
-        self.assertEqual(readthrough.start_date.month, 4)
-        self.assertEqual(readthrough.start_date.day, 16)
-        self.assertEqual(readthrough.finish_date.year, 2007)
-        self.assertEqual(readthrough.finish_date.month, 5)
-        self.assertEqual(readthrough.finish_date.day, 8)
+        self.assertEqual(readthrough.start_date.year, 2020)
+        self.assertEqual(readthrough.start_date.month, 10)
+        self.assertEqual(readthrough.start_date.day, 21)
+        self.assertEqual(readthrough.finish_date.year, 2020)
+        self.assertEqual(readthrough.finish_date.month, 10)
+        self.assertEqual(readthrough.finish_date.day, 25)
 
     def test_handle_import_twice(self):
         """ re-importing books """
         shelf = self.user.shelf_set.filter(identifier="read").first()
         import_job = models.ImportJob.objects.create(user=self.user)
-        datafile = pathlib.Path(__file__).parent.joinpath("data/librarything.tsv")
-        csv_file = open(datafile, "r", encoding=self.importer.encoding)
-        for index, entry in enumerate(
-            list(csv.DictReader(csv_file, delimiter=self.importer.delimiter))
-        ):
+        datafile = pathlib.Path(__file__).parent.joinpath("data/goodreads.csv")
+        csv_file = open(datafile, "r")
+        for index, entry in enumerate(list(csv.DictReader(csv_file))):
             entry = self.importer.parse_fields(entry)
             import_item = models.ImportItem.objects.create(
                 job_id=import_job.id, index=index, data=entry, book=self.book
@@ -192,20 +196,20 @@ class LibrarythingImport(TestCase):
         readthrough = models.ReadThrough.objects.get(user=self.user)
         self.assertEqual(readthrough.book, self.book)
         # I can't remember how to create dates and I don't want to look it up.
-        self.assertEqual(readthrough.start_date.year, 2007)
-        self.assertEqual(readthrough.start_date.month, 4)
-        self.assertEqual(readthrough.start_date.day, 16)
-        self.assertEqual(readthrough.finish_date.year, 2007)
-        self.assertEqual(readthrough.finish_date.month, 5)
-        self.assertEqual(readthrough.finish_date.day, 8)
+        self.assertEqual(readthrough.start_date.year, 2020)
+        self.assertEqual(readthrough.start_date.month, 10)
+        self.assertEqual(readthrough.start_date.day, 21)
+        self.assertEqual(readthrough.finish_date.year, 2020)
+        self.assertEqual(readthrough.finish_date.month, 10)
+        self.assertEqual(readthrough.finish_date.day, 25)
 
     @patch("bookwyrm.activitystreams.ActivityStream.add_status")
     def test_handle_imported_book_review(self, _):
-        """ librarything review import """
+        """ goodreads review import """
         import_job = models.ImportJob.objects.create(user=self.user)
-        datafile = pathlib.Path(__file__).parent.joinpath("data/librarything.tsv")
-        csv_file = open(datafile, "r", encoding=self.importer.encoding)
-        entry = list(csv.DictReader(csv_file, delimiter=self.importer.delimiter))[0]
+        datafile = pathlib.Path(__file__).parent.joinpath("data/goodreads.csv")
+        csv_file = open(datafile, "r")
+        entry = list(csv.DictReader(csv_file))[2]
         entry = self.importer.parse_fields(entry)
         import_item = models.ImportItem.objects.create(
             job_id=import_job.id, index=0, data=entry, book=self.book
@@ -216,19 +220,19 @@ class LibrarythingImport(TestCase):
                 self.importer.service, self.user, import_item, True, "unlisted"
             )
         review = models.Review.objects.get(book=self.book, user=self.user)
-        self.assertEqual(review.content, "chef d'oeuvre")
-        self.assertEqual(review.rating, 5)
-        self.assertEqual(review.published_date.year, 2007)
-        self.assertEqual(review.published_date.month, 5)
+        self.assertEqual(review.content, "mixed feelings")
+        self.assertEqual(review.rating, 2)
+        self.assertEqual(review.published_date.year, 2019)
+        self.assertEqual(review.published_date.month, 7)
         self.assertEqual(review.published_date.day, 8)
         self.assertEqual(review.privacy, "unlisted")
 
     def test_handle_imported_book_reviews_disabled(self):
-        """ librarything review import """
+        """ goodreads review import """
         import_job = models.ImportJob.objects.create(user=self.user)
-        datafile = pathlib.Path(__file__).parent.joinpath("data/librarything.tsv")
-        csv_file = open(datafile, "r", encoding=self.importer.encoding)
-        entry = list(csv.DictReader(csv_file, delimiter=self.importer.delimiter))[2]
+        datafile = pathlib.Path(__file__).parent.joinpath("data/goodreads.csv")
+        csv_file = open(datafile, "r")
+        entry = list(csv.DictReader(csv_file))[2]
         entry = self.importer.parse_fields(entry)
         import_item = models.ImportItem.objects.create(
             job_id=import_job.id, index=0, data=entry, book=self.book
