+++ bookwyrm/tests/importers/test_librarything_import.py
@@ -1,5 +1,4 @@
 """ testing import """
-from collections import namedtuple
 import csv
 import pathlib
 from unittest.mock import patch
@@ -7,21 +6,23 @@ from unittest.mock import patch
 from django.test import TestCase
 import responses
 
-from bookwyrm import models, importer
-from bookwyrm.goodreads_import import GoodreadsImporter
+from bookwyrm import models
+from bookwyrm.importers import importer, LibrarythingImporter
 from bookwyrm.settings import DOMAIN
 
 
-class GoodreadsImport(TestCase):
-    """ importing from goodreads csv """
+class LibrarythingImport(TestCase):
+    """ importing from librarything tsv """
 
     def setUp(self):
-        """ use a test csv """
-        self.importer = GoodreadsImporter()
-        datafile = pathlib.Path(__file__).parent.joinpath("data/goodreads.csv")
+        """ use a test tsv """
+        self.importer = LibrarythingImporter()
+        datafile = pathlib.Path(__file__).parent.joinpath("data/librarything.tsv")
+
+        # Librarything generates latin encoded exports...
         self.csv = open(datafile, "r", encoding=self.importer.encoding)
         self.user = models.User.objects.create_user(
-            "mouse", "mouse@mouse.mouse", "password", local=True
+            "mmai", "mmai@mmai.mmai", "password", local=True
         )
 
         models.Connector.objects.create(
@@ -52,11 +53,11 @@ class GoodreadsImport(TestCase):
         import_items = models.ImportItem.objects.filter(job=import_job).all()
         self.assertEqual(len(import_items), 3)
         self.assertEqual(import_items[0].index, 0)
-        self.assertEqual(import_items[0].data["Book Id"], "42036538")
+        self.assertEqual(import_items[0].data["Book Id"], "5498194")
         self.assertEqual(import_items[1].index, 1)
-        self.assertEqual(import_items[1].data["Book Id"], "52691223")
+        self.assertEqual(import_items[1].data["Book Id"], "5015319")
         self.assertEqual(import_items[2].index, 2)
-        self.assertEqual(import_items[2].data["Book Id"], "28694510")
+        self.assertEqual(import_items[2].data["Book Id"], "5015399")
 
     def test_create_retry_job(self):
         """ trying again with items that didn't import """
@@ -72,20 +73,9 @@ class GoodreadsImport(TestCase):
         retry_items = models.ImportItem.objects.filter(job=retry).all()
         self.assertEqual(len(retry_items), 2)
         self.assertEqual(retry_items[0].index, 0)
-        self.assertEqual(retry_items[0].data["Book Id"], "42036538")
+        self.assertEqual(import_items[0].data["Book Id"], "5498194")
         self.assertEqual(retry_items[1].index, 1)
-        self.assertEqual(retry_items[1].data["Book Id"], "52691223")
-
-    def test_start_import(self):
-        """ begin loading books """
-        import_job = self.importer.create_job(self.user, self.csv, False, "unlisted")
-        MockTask = namedtuple("Task", ("id"))
-        mock_task = MockTask(7)
-        with patch("bookwyrm.importer.import_data.delay") as start:
-            start.return_value = mock_task
-            self.importer.start_import(import_job)
-        import_job.refresh_from_db()
-        self.assertEqual(import_job.task_id, "7")
+        self.assertEqual(retry_items[1].data["Book Id"], "5015319")
 
     @responses.activate
     def test_import_data(self):
@@ -104,14 +94,16 @@ class GoodreadsImport(TestCase):
         self.assertEqual(import_item.book.id, book.id)
 
     def test_handle_imported_book(self):
-        """ goodreads import added a book, this adds related connections """
+        """ librarything import added a book, this adds related connections """
         shelf = self.user.shelf_set.filter(identifier="read").first()
         self.assertIsNone(shelf.books.first())
 
         import_job = models.ImportJob.objects.create(user=self.user)
-        datafile = pathlib.Path(__file__).parent.joinpath("data/goodreads.csv")
-        csv_file = open(datafile, "r")
-        for index, entry in enumerate(list(csv.DictReader(csv_file))):
+        datafile = pathlib.Path(__file__).parent.joinpath("data/librarything.tsv")
+        csv_file = open(datafile, "r", encoding=self.importer.encoding)
+        for index, entry in enumerate(
+            list(csv.DictReader(csv_file, delimiter=self.importer.delimiter))
+        ):
             entry = self.importer.parse_fields(entry)
             import_item = models.ImportItem.objects.create(
                 job_id=import_job.id, index=index, data=entry, book=self.book
@@ -129,23 +121,25 @@ class GoodreadsImport(TestCase):
         readthrough = models.ReadThrough.objects.get(user=self.user)
         self.assertEqual(readthrough.book, self.book)
         # I can't remember how to create dates and I don't want to look it up.
-        self.assertEqual(readthrough.start_date.year, 2020)
-        self.assertEqual(readthrough.start_date.month, 10)
-        self.assertEqual(readthrough.start_date.day, 21)
-        self.assertEqual(readthrough.finish_date.year, 2020)
-        self.assertEqual(readthrough.finish_date.month, 10)
-        self.assertEqual(readthrough.finish_date.day, 25)
+        self.assertEqual(readthrough.start_date.year, 2007)
+        self.assertEqual(readthrough.start_date.month, 4)
+        self.assertEqual(readthrough.start_date.day, 16)
+        self.assertEqual(readthrough.finish_date.year, 2007)
+        self.assertEqual(readthrough.finish_date.month, 5)
+        self.assertEqual(readthrough.finish_date.day, 8)
 
     def test_handle_imported_book_already_shelved(self):
-        """ goodreads import added a book, this adds related connections """
+        """ librarything import added a book, this adds related connections """
         with patch("bookwyrm.models.activitypub_mixin.broadcast_task.delay"):
             shelf = self.user.shelf_set.filter(identifier="to-read").first()
             models.ShelfBook.objects.create(shelf=shelf, user=self.user, book=self.book)
 
         import_job = models.ImportJob.objects.create(user=self.user)
-        datafile = pathlib.Path(__file__).parent.joinpath("data/goodreads.csv")
-        csv_file = open(datafile, "r")
-        for index, entry in enumerate(list(csv.DictReader(csv_file))):
+        datafile = pathlib.Path(__file__).parent.joinpath("data/librarything.tsv")
+        csv_file = open(datafile, "r", encoding=self.importer.encoding)
+        for index, entry in enumerate(
+            list(csv.DictReader(csv_file, delimiter=self.importer.delimiter))
+        ):
             entry = self.importer.parse_fields(entry)
             import_item = models.ImportItem.objects.create(
                 job_id=import_job.id, index=index, data=entry, book=self.book
@@ -162,20 +156,22 @@ class GoodreadsImport(TestCase):
         self.assertIsNone(self.user.shelf_set.get(identifier="read").books.first())
         readthrough = models.ReadThrough.objects.get(user=self.user)
         self.assertEqual(readthrough.book, self.book)
-        self.assertEqual(readthrough.start_date.year, 2020)
-        self.assertEqual(readthrough.start_date.month, 10)
-        self.assertEqual(readthrough.start_date.day, 21)
-        self.assertEqual(readthrough.finish_date.year, 2020)
-        self.assertEqual(readthrough.finish_date.month, 10)
-        self.assertEqual(readthrough.finish_date.day, 25)
+        self.assertEqual(readthrough.start_date.year, 2007)
+        self.assertEqual(readthrough.start_date.month, 4)
+        self.assertEqual(readthrough.start_date.day, 16)
+        self.assertEqual(readthrough.finish_date.year, 2007)
+        self.assertEqual(readthrough.finish_date.month, 5)
+        self.assertEqual(readthrough.finish_date.day, 8)
 
     def test_handle_import_twice(self):
         """ re-importing books """
         shelf = self.user.shelf_set.filter(identifier="read").first()
         import_job = models.ImportJob.objects.create(user=self.user)
-        datafile = pathlib.Path(__file__).parent.joinpath("data/goodreads.csv")
-        csv_file = open(datafile, "r")
-        for index, entry in enumerate(list(csv.DictReader(csv_file))):
+        datafile = pathlib.Path(__file__).parent.joinpath("data/librarything.tsv")
+        csv_file = open(datafile, "r", encoding=self.importer.encoding)
+        for index, entry in enumerate(
+            list(csv.DictReader(csv_file, delimiter=self.importer.delimiter))
+        ):
             entry = self.importer.parse_fields(entry)
             import_item = models.ImportItem.objects.create(
                 job_id=import_job.id, index=index, data=entry, book=self.book
@@ -196,20 +192,20 @@ class GoodreadsImport(TestCase):
         readthrough = models.ReadThrough.objects.get(user=self.user)
         self.assertEqual(readthrough.book, self.book)
         # I can't remember how to create dates and I don't want to look it up.
-        self.assertEqual(readthrough.start_date.year, 2020)
-        self.assertEqual(readthrough.start_date.month, 10)
-        self.assertEqual(readthrough.start_date.day, 21)
-        self.assertEqual(readthrough.finish_date.year, 2020)
-        self.assertEqual(readthrough.finish_date.month, 10)
-        self.assertEqual(readthrough.finish_date.day, 25)
+        self.assertEqual(readthrough.start_date.year, 2007)
+        self.assertEqual(readthrough.start_date.month, 4)
+        self.assertEqual(readthrough.start_date.day, 16)
+        self.assertEqual(readthrough.finish_date.year, 2007)
+        self.assertEqual(readthrough.finish_date.month, 5)
+        self.assertEqual(readthrough.finish_date.day, 8)
 
     @patch("bookwyrm.activitystreams.ActivityStream.add_status")
     def test_handle_imported_book_review(self, _):
-        """ goodreads review import """
+        """ librarything review import """
         import_job = models.ImportJob.objects.create(user=self.user)
-        datafile = pathlib.Path(__file__).parent.joinpath("data/goodreads.csv")
-        csv_file = open(datafile, "r")
-        entry = list(csv.DictReader(csv_file))[2]
+        datafile = pathlib.Path(__file__).parent.joinpath("data/librarything.tsv")
+        csv_file = open(datafile, "r", encoding=self.importer.encoding)
+        entry = list(csv.DictReader(csv_file, delimiter=self.importer.delimiter))[0]
         entry = self.importer.parse_fields(entry)
         import_item = models.ImportItem.objects.create(
             job_id=import_job.id, index=0, data=entry, book=self.book
@@ -220,19 +216,19 @@ class GoodreadsImport(TestCase):
                 self.importer.service, self.user, import_item, True, "unlisted"
             )
         review = models.Review.objects.get(book=self.book, user=self.user)
-        self.assertEqual(review.content, "mixed feelings")
-        self.assertEqual(review.rating, 2)
-        self.assertEqual(review.published_date.year, 2019)
-        self.assertEqual(review.published_date.month, 7)
+        self.assertEqual(review.content, "chef d'oeuvre")
+        self.assertEqual(review.rating, 5)
+        self.assertEqual(review.published_date.year, 2007)
+        self.assertEqual(review.published_date.month, 5)
         self.assertEqual(review.published_date.day, 8)
         self.assertEqual(review.privacy, "unlisted")
 
     def test_handle_imported_book_reviews_disabled(self):
-        """ goodreads review import """
+        """ librarything review import """
         import_job = models.ImportJob.objects.create(user=self.user)
-        datafile = pathlib.Path(__file__).parent.joinpath("data/goodreads.csv")
-        csv_file = open(datafile, "r")
-        entry = list(csv.DictReader(csv_file))[2]
+        datafile = pathlib.Path(__file__).parent.joinpath("data/librarything.tsv")
+        csv_file = open(datafile, "r", encoding=self.importer.encoding)
+        entry = list(csv.DictReader(csv_file, delimiter=self.importer.delimiter))[2]
         entry = self.importer.parse_fields(entry)
         import_item = models.ImportItem.objects.create(
             job_id=import_job.id, index=0, data=entry, book=self.book
